Bring ADC3 online - L2, R2, Battery
This changeset brings ADC3 online tentatively with the same oversampling settings as used with the Joystick.

Currently the same deadzone is applied to L2 and R2 but since these are unlikely to be biased to the center (albeit who knows what the enterprising 32blit hacker might do) we might have to reconsider how this works, or just provide a raw (but oversampled and stabalised) ADC reading.

The battery readout is stored in blit::battery, this needs moving. The ADC is 16bit and the voltage can be calculated as: 6.6 / 65535 * ADC_READING.

Battery voltage should range from 4.2V down to 3V at which point we should add a soft shutdown into shipping mode. The hardware safety cutoff is around 2.4V.

Discharge rate is non-linear, we should measure it to get a feel for how to scale voltages to better represent remaining battery life for 32blit.