Change "fire" to "rotozoom"
Since Doom Fire is a much better example of an on-device fire effect, and "fire" was actually showing a rotozoom.

I've also fixed a bug whereby the code to switch between lores/hires mode on the fly was causing flickering.

Basically, this turns the mess that was "fire" into a worthwhile example showing a rotozoom demo effect.