Fixed bugs in scrolly-tile
In addition to fixing (I hope) the bugs that caused abrupt endings to passages I've deleted some dead code and improved code re-use.

I've also added little pools of water in the bottom of emergent passages. Looks kinda naff for now.

Converting this to high-res might be nice- it's hard to make a character sprite out of a handful of pixels. Althought the simple box is kinda endearing.

I'd also love to have a small back-buffer allowing you to fall back down passages a little way to explore- easy to make it work, but not sure how it'd feel mechanically.